Windows service support initial changes

+++ include/dns/dnscommon.hpp
@@ -127,7 +127,7 @@ enum PacketType {
 struct BytePacketBuffer {
   std::array<uint8_t, MAX_PACKET_SZ> buf{};
   uint16_t pos{0};
-  uint16_t size{0};
+  size_t size{0};
 };
 
 class DnsObject {
